requirements.txt
Flask == 3.0.3
Werkzeug == 3.0.2
Jinja2 == 3.1.3
Python-dotenv == 1.0.1
Flask-Mail==0.9.1